requirements.txt
celery==5.2.1
Django==3.2.10
django-celery-beat==2.2.1
django-celery-results==2.2.0
djangorestframework==3.13.1
python-dotenv==0.19.2
redis==4.0.2
requests==2.26.0
uuid==1.30